requirements.txt
pkg-resources==0.0.0
RPi.GPIO==0.7.0
RPLCD==1.2.2
